pyproject.toml
[project]
name = "labscare-script-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-docx>=1.1.0",
]

[project.scripts]
get-docx-comment = "get_docx_comment:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
